backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
curl-cffi==0.7.1
beautifulsoup4==4.12.3
lxml==5.2.2
apscheduler==3.10.4
python-dotenv==1.0.1
aiosmtplib==3.0.1
jinja2==3.1.4
pydantic==2.7.1
